requirements.txt
python-magic

atomicwrites==1.2.0
attrs==18.1.0
astroid==2.0.4
coverage==4.5.1
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-extensions==2.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
facebook-sdk==3.0.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.9.0
more-itertools==4.3.0
nose==1.3.7
pluggy==0.7.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pylint==2.1.1
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
wrapt==1.10.11
